kvdb related changes
Addresses issue https://github.com/bombardier-gif/covid-vaccine-booking/issues/417 using a binary for windows user. On this error user is expected to provide new bucket id and use it. It asks for it once and then stores it in json file and reads from there on subsequent runs. If json is not stored, user will have to provide it manually every time they run the exe.